save torch_npu modifications

+++ deepmd/env.py
@@ -33,12 +33,12 @@ CONFIG_FILE = SHARED_LIB_DIR / "run_config.ini"
 dp_float_prec = os.environ.get("DP_INTERFACE_PREC", "high").lower()
 if dp_float_prec in ("high", ""):
     # default is high
-    GLOBAL_NP_FLOAT_PRECISION = np.float64
-    GLOBAL_ENER_FLOAT_PRECISION = np.float64
+    GLOBAL_NP_FLOAT_PRECISION = np.float32
+    GLOBAL_ENER_FLOAT_PRECISION = np.float32
     global_float_prec = "double"
 elif dp_float_prec == "low":
-    GLOBAL_NP_FLOAT_PRECISION = np.float32
-    GLOBAL_ENER_FLOAT_PRECISION = np.float64
+    GLOBAL_NP_FLOAT_PRECISION = np.float16
+    GLOBAL_ENER_FLOAT_PRECISION = np.float16
     global_float_prec = "float"
 else:
     raise RuntimeError(
